api/focus: Another set of rearrangements
Instead of `colormap` & co importing focus and registering themselves, let focus
import and register them. This lets us avoid a dependency loop, and can simplify
our layout further.
